refactor(room): split shared.ts into protocol / state / participant
One file carried three abstractions. Now each is its own, named for what
it is:

- protocol.ts — everything both sides must agree on and nothing that
  behaves: keys, records, wire envelopes, binary framing, want shapes.
  (Framing now precedes the wants that reference its constants — the
  file reads top-to-bottom.)
- state.ts — RoomState, the event-driven local view shared by server
  and client, plus the RemoteParticipant backing brand.
- participant.ts — ParticipantBase, the shared LocalParticipant half
  (inbox hold + leave lifecycle).

Pure moves; importers updated. No behavior change.
