--- packages/telefunc/wire-protocol/room/shared.ts
@@ -1,1231 +1,0 @@
-export {
-  ROOM_KEY_NAMESPACE,
-  roomCtrlKey,
-  roomTextKey,
-  roomMemberDataKey,
-  roomMemberTrackKey,
-  roomActivityKey,
-  roomDmKey,
-  roomConfigKvKey,
-  roomIdFromConfigKey,
-  roomMemberKvKey,
-  roomMemberKvPrefix,
-  sizeToWire,
-  sizeFromWire,
-  uuidToBytes,
-  frameWithMemberId,
-  unframeMemberId,
-  hasRoomTag,
-  normalizeJoinOptions,
-  stampNewer,
-  errorMessage,
-  leaveCauseFromWire,
-  leaveCauseToWire,
-  remoteBacking,
-  DEFAULT_TRACK,
-  SUB_BINARY_MEMBERS_MAX,
-  SUB_BINARY_TRACKS_MAX,
-  emptyTrackWants,
-  mergeTrackWants,
-  wantsTrack,
-  wantsAnyBinary,
-  sanitizeBinaryWants,
-  RoomState,
-  ParticipantBase,
-}
-export type {
-  RoomConfigRecord,
-  RoomMemberRecord,
-  MemberSnapshot,
-  RoomSnapshotMetadata,
-  ParticipantStubMetadata,
-  RoomEnvelope,
-  RoomRosterEvent,
-  RoomCtrlEnvelope,
-  RoomDataEnvelope,
-  RoomDataPublish,
-  RoomAnnounceEnvelope,
-  RoomDmEnvelope,
-  RoomStubRequest,
-  ParticipantStubRequest,
-  ParticipantStubNotice,
-  ReqOkAck,
-  ReqJoinAck,
-  ReqPublishAck,
-  MemberWants,
-  TrackWants,
-  BinaryWants,
-  RoomActivityEvent,
-  InboxMessage,
-}
-
-import { assert, assertUsage } from '../../utils/assert.js'
-import { isObject } from '../../utils/isObject.js'
-import type { ChannelPublishAck, ChannelPublishInfo } from '../channel.js'
-import type {
-  BinaryFrameInfo,
-  BinaryPublishOptions,
-  RoomSnapshotView,
-  JoinOptions,
-  LeaveCause,
-  LocalParticipant,
-  ParticipantMeta,
-  RemoteParticipant,
-  RoomMeta,
-  Sender,
-} from './types.js'
-
-// ---------------------------------------------------------------------------
-// Keys & records
-// ---------------------------------------------------------------------------
-
-/** Reserved pub/sub + KV namespace for rooms. Don't use it for `BroadcastChannel` keys. */
-const ROOM_KEY_NAMESPACE = 'telefunc:room:'
-
-/** Pub/sub key carrying a room's control lane: presence & lifecycle events plus room-authored
- *  announcements. Low-rate, subscribed by every observer — never carries participant data. */
-function roomCtrlKey(roomId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}`
-}
-/** Pub/sub key carrying the room's text data in shared mode — its own lane, so holders that
- *  only observe presence never receive it. */
-function roomTextKey(roomId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:t`
-}
-/** Pub/sub key carrying one member's data: default-track binary always (per-publisher keys
- *  make delivery member-selective at the source), text too in isolated mode. */
-function roomMemberDataKey(roomId: string, memberId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:m:${memberId}`
-}
-/** Pub/sub key carrying one member's named binary track. Per-(member, track) keys make delivery
- *  track-selective at the source: a holder that stops watching a track drops this subscription,
- *  and the publisher's `receivers` hits 0 when nobody anywhere holds it — bytes stop flowing at
- *  every hop, not just at delivery. */
-function roomMemberTrackKey(roomId: string, memberId: string, track: string): string {
-  return `${roomMemberDataKey(roomId, memberId)}:t:${track}`
-}
-/** Pub/sub key carrying the room's throttled activity signal — subscribed only by holders with
- *  `onActivity` listeners (badge consumers), so nobody else pays even that trickle. */
-function roomActivityKey(roomId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:a`
-}
-/** Pub/sub key carrying one member's private inbox — only the member's owning node subscribes. */
-function roomDmKey(roomId: string, memberId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:dm:${memberId}`
-}
-/** KV key of the room's config record. */
-function roomConfigKvKey(roomId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:config`
-}
-/** Inverse of `roomConfigKvKey` — `null` for keys that aren't room config records. */
-function roomIdFromConfigKey(key: string): string | null {
-  if (!key.startsWith(ROOM_KEY_NAMESPACE) || !key.endsWith(':config')) return null
-  return key.slice(ROOM_KEY_NAMESPACE.length, -':config'.length)
-}
-/** KV key of one member record. */
-function roomMemberKvKey(roomId: string, memberId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:m:${memberId}`
-}
-/** KV prefix under which all of a room's member records live. Member IDs are UUIDs, which is
- *  how member records are told apart from keys of other rooms whose ID shares the prefix. */
-function roomMemberKvPrefix(roomId: string): string {
-  return `${ROOM_KEY_NAMESPACE}${roomId}:m:`
-}
-
-/** Stored at `roomConfigKvKey`. `size: null` encodes `Infinity` (not JSON-safe). `at`/`by` is
- *  the last-writer-wins stamp of the latest `Room.update()` (see `applyRoomUpdate`). */
-type RoomConfigRecord = {
-  meta: RoomMeta
-  size: number | null
-  isolated: boolean
-  at: number
-  by: string
-}
-
-/** Later timestamp wins; equal timestamps break deterministically by writer ID. */
-function stampNewer(a: { at: number; by: string }, b: { at: number; by: string }): boolean {
-  return a.at > b.at || (a.at === b.at && a.by > b.by)
-}
-
-/** Stored at `roomMemberKvKey`. `seenAt` is the liveness timestamp: the owning node refreshes
- *  it every `ROOM_HEARTBEAT_INTERVAL_MS`; records older than `ROOM_MEMBER_TTL_MS` are reaped. */
-type RoomMemberRecord = {
-  meta: ParticipantMeta
-  joinedAt: number
-  seenAt: number
-  /** Monotonic meta revision, issued by the member's single owner — orders `p-meta` events. */
-  metaSeq: number
-  /** App identity stamped at (server-side) join — absent: none. Immutable per member. */
-  identity?: string
-  /** Named binary tracks this member has published — appended by the owner before the first
-   *  frame of each track, so late observers can subscribe every track they can't name. */
-  tracks?: string[]
-}
-
-function sizeToWire(size: number): number | null {
-  return Number.isFinite(size) ? size : null
-}
-function sizeFromWire(size: number | null): number {
-  return size ?? Infinity
-}
-
-// ---------------------------------------------------------------------------
-// Wire shapes
-// ---------------------------------------------------------------------------
-
-type MemberSnapshot = {
-  id: string
-  meta: ParticipantMeta
-  joinedAt: number
-  metaSeq: number
-  /** App identity stamped at (server-side) join — `null`/absent: none. Immutable per member. */
-  identity?: string | null
-  /** Named binary tracks the member has published (see `RoomMemberRecord.tracks`). */
-  tracks?: string[]
-}
-
-/** Serializer metadata of a `Room` crossing the wire. Carries only scalars — the roster itself
- *  streams as a `RoomRosterEvent` once the stub's peer attaches, so serialization stays O(1)
- *  no matter how many members the room has. */
-type RoomSnapshotMetadata = {
-  channelId: string
-  roomId: string
-  meta: RoomMeta
-  size: number | null
-  isolated: boolean
-  closed: boolean
-  count: number
-  /** LWW stamp of the config the snapshot reflects — seeds `applyRoomUpdate` ordering. */
-  stamp: { at: number; by: string }
-}
-
-/** Serializer metadata of a `LocalParticipant` crossing the wire. */
-type ParticipantStubMetadata = {
-  channelId: string
-  roomId: string
-  id: string
-  meta: ParticipantMeta
-  joinedAt: number
-  selfDelivery: boolean
-  identity: string | null
-}
-
-/** Presence & lifecycle events, published on the room's control key by whichever node caused them.
- *
- *  The origin applies its own event locally (for deterministic same-node semantics) and then
- *  receives it back via the pub/sub echo. `join`/`leave`/`closed` are naturally idempotent;
- *  `p-meta` orders by the owner-issued `seq`, `update` by its `at`/`by` stamp — echoes and
- *  concurrent writers converge to the same winner on every node, whatever the arrival order. */
-type RoomCtrlEnvelope =
-  | { __r: 'join'; id: string; meta: ParticipantMeta; joinedAt: number; identity?: string }
-  | { __r: 'leave'; id: string; cause?: 'removed' | 'disconnected'; reason?: unknown }
-  | { __r: 'p-meta'; id: string; meta: ParticipantMeta; prev: ParticipantMeta; seq: number }
-  | { __r: 'update'; meta: RoomMeta; prev: RoomMeta; size: number | null; at: number; by: string }
-  // A member's first publish on a new named track — announced before the frame, so live
-  // all-track subscribers bring up the track-key subscription (idempotent, like join).
-  | { __r: 'track'; id: string; track: string }
-  | { __r: 'closed' }
-
-/** A participant's message. Published on the room's text key (shared mode) or the member's own
- *  key (isolated mode). `fromMeta` is the sender's meta as verified by the sender's own node —
- *  never client-supplied — so any receiver can surface a correct sender even before its roster
- *  view catches up (see `RoomState.applyData`). */
-type RoomDataEnvelope = { __r: 'data'; from: string; fromMeta: ParticipantMeta; fromIdentity?: string; data: unknown }
-
-/** What a client sends upward to publish — its node verifies membership and stamps `fromMeta`. */
-type RoomDataPublish = { __r: 'data'; from: string; data: unknown }
-
-/** The activity trickle (own key `:a`): "something was published around `at`" — throttled at
- *  the publisher, consumed by badge-style listeners that don't want message bodies. */
-type RoomActivityEvent = { __r: 'activity'; at: number }
-
-/** A room-authored message (`Room.announce()`) — no sender, delivered to `onAnnounce()`. */
-type RoomAnnounceEnvelope = { __r: 'announce'; data: unknown }
-
-type RoomEnvelope = RoomCtrlEnvelope | RoomDataEnvelope | RoomAnnounceEnvelope
-
-/** The authoritative roster, pushed once per stub after its peer attaches — never published on
- *  the adapter. Position-in-stream consistency: every event relayed before it is already
- *  reflected in it; later events apply incrementally on top. */
-type RoomRosterEvent = { __r: 'roster'; members: MemberSnapshot[] }
-
-/** A direct message, published on the target's inbox key (`roomDmKey`) — transport-level
- *  privacy: only the target's owning node subscribes, only its holder receives the relay.
- *  `to` lets a holder of several participants route the message to the right one. */
-type RoomDmEnvelope = {
-  __r: 'dm'
-  to: string
-  from: string
-  fromMeta: ParticipantMeta | null
-  fromIdentity?: string
-  data: unknown
-}
-
-/** Client→server requests on a `Room` stub channel. `id` identifies the sending participant.
- *  `sub-binary` declares the client's binary wants (full replace, see `BinaryWants`);
- *  `sub-text` declares member-scoped text wants — the room-level (all) text want rides the
- *  standard broadcast-subscription ctrl instead, keeping its synchronous-declaration fence. */
-type RoomStubRequest =
-  | { __r: 'req-join'; meta: ParticipantMeta; selfDelivery: boolean }
-  | { __r: 'req-leave'; id: string }
-  | { __r: 'req-set-meta'; id: string; meta: ParticipantMeta }
-  | { __r: 'req-dm'; id: string; to: string; data: unknown }
-  | { __r: 'sub-binary'; wants: BinaryWants }
-  | { __r: 'sub-text'; members: string[] }
-  | { __r: 'sub-activity'; on: boolean }
-
-/** Client→server requests on a standalone `LocalParticipant` stub channel. */
-type ParticipantStubRequest =
-  | { __r: 'req-publish'; data: unknown }
-  | { __r: 'req-set-meta'; meta: ParticipantMeta }
-  | { __r: 'req-dm'; to: string; data: unknown }
-  | { __r: 'req-leave' }
-
-/** Server→client notices on a standalone `LocalParticipant` stub channel. */
-type ParticipantStubNotice =
-  | { __r: 'left'; cause?: 'removed' | 'disconnected' | 'closed'; reason?: unknown }
-  | { __r: 'p-meta'; meta: ParticipantMeta }
-  | { __r: 'dm'; from: string; fromMeta: ParticipantMeta | null; fromIdentity?: string; data: unknown }
-
-/** Which members' streams a holder wants on the text lane — `all` for room-level listeners,
- *  or a specific member set for participant-scoped ones. */
-type MemberWants = { all: boolean; members: string[] }
-
-// ---------------------------------------------------------------------------
-// Binary wants — per member, per track
-// ---------------------------------------------------------------------------
-
-/** The default (unnamed) track's slot in want sets and key routing — track names are non-empty
- *  by contract (`frameWithMemberId`), so `''` is unambiguous. */
-const DEFAULT_TRACK = ''
-
-/** Which of a publisher's tracks a holder wants: every track, or an exact set
- *  (`DEFAULT_TRACK` selects the unnamed lane). */
-type TrackWants = { all: boolean; tracks: string[] }
-
-/** A holder's complete binary wants. `everyMember` comes from room-level listeners and applies
- *  to all members; `members` adds participant-scoped wants on top. This one shape drives all
- *  three gates: the client's declaration, the server's upstream key set, and the per-stub relay. */
-type BinaryWants = { everyMember: TrackWants; members: Record<string, TrackWants> }
-
-/** Abuse bounds on a client's `sub-binary` declaration — generous for real apps (a media app
- *  uses a handful of tracks), fatal for hostile blowups (tracks multiply upstream keys). */
-const SUB_BINARY_MEMBERS_MAX = 4096
-const SUB_BINARY_TRACKS_MAX = 64
-
-function emptyTrackWants(): TrackWants {
-  return { all: false, tracks: [] }
-}
-
-function mergeTrackWants(a: TrackWants, b: TrackWants): TrackWants {
-  if (a.all || b.all) return { all: true, tracks: [] }
-  return { all: false, tracks: [...new Set([...a.tracks, ...b.tracks])] }
-}
-
-function wantsTrack(wants: TrackWants, track: string): boolean {
-  return wants.all || wants.tracks.includes(track)
-}
-
-function wantsAnyBinary(wants: BinaryWants): boolean {
-  return wants.everyMember.all || wants.everyMember.tracks.length > 0 || Object.keys(wants.members).length > 0
-}
-
-/** Validate a client-declared `sub-binary` want (untrusted input) — bounded and well-formed,
- *  or `null` to reject the declaration. */
-function sanitizeBinaryWants(wants: unknown): BinaryWants | null {
-  if (!isObject(wants)) return null
-  const everyMember = sanitizeTrackWants(wants.everyMember)
-  if (!everyMember || !isObject(wants.members)) return null
-  const members: Record<string, TrackWants> = {}
-  const entries = Object.entries(wants.members)
-  if (entries.length > SUB_BINARY_MEMBERS_MAX) return null
-  for (const [memberId, trackWants] of entries) {
-    const sanitized = sanitizeTrackWants(trackWants)
-    if (!sanitized) return null
-    members[memberId] = sanitized
-  }
-  return { everyMember, members }
-}
-
-function sanitizeTrackWants(wants: unknown): TrackWants | null {
-  if (!isObject(wants) || typeof wants.all !== 'boolean' || !Array.isArray(wants.tracks)) return null
-  if (wants.tracks.length > SUB_BINARY_TRACKS_MAX) return null
-  if (!wants.tracks.every((track) => typeof track === 'string' && track.length <= TRACK_MAX_BYTES)) return null
-  return { all: wants.all, tracks: wants.tracks as string[] }
-}
-
-type ReqOkAck = { ok: true } | { ok: false; err: string }
-type ReqJoinAck = { ok: true; id: string; joinedAt: number } | { ok: false; err: string }
-type ReqPublishAck = { ok: true; ack: ChannelPublishInfo } | { ok: false; err: string }
-
-/** Decode a leave event's cause — an absent wire cause means a voluntary leave. */
-function leaveCauseFromWire(event: { cause?: 'removed' | 'disconnected'; reason?: unknown }): LeaveCause {
-  if (event.cause === 'removed') {
-    return event.reason === undefined ? { type: 'removed' } : { type: 'removed', reason: event.reason }
-  }
-  return { type: event.cause === 'disconnected' ? 'disconnected' : 'left' }
-}
-
-/** Encode a cause into leave-event fields — `'left'` is the wire default and travels as nothing. */
-function leaveCauseToWire(cause: LeaveCause): { cause?: 'removed' | 'disconnected'; reason?: unknown } {
-  if (cause.type === 'removed')
-    return cause.reason === undefined ? { cause: 'removed' } : { cause: 'removed', reason: cause.reason }
-  if (cause.type === 'disconnected') return { cause: 'disconnected' }
-  return {}
-}
-
-/** All room messages are tagged with `__r` — envelopes, requests, and notices alike. */
-function hasRoomTag(value: unknown): value is { __r: string } {
-  return isObject(value) && typeof value.__r === 'string'
-}
-
-/** Validates `join(meta, options)` arguments; returns the resolved `selfDelivery`. */
-function normalizeJoinOptions(meta: unknown, options: JoinOptions | undefined): boolean {
-  assertUsage(isObject(meta), 'join() meta should be an object')
-  assertUsage(options === undefined || isObject(options), 'join() options should be an object')
-  return options?.selfDelivery !== false
-}
-
-/** How errors travel inside `ok: false` acks. */
-function errorMessage(err: unknown): string {
-  return err instanceof Error ? err.message : String(err)
-}
-
-// ---------------------------------------------------------------------------
-// Member IDs — UUIDs, framed as a fixed 16-byte prefix on binary messages
-// ---------------------------------------------------------------------------
-
-const MEMBER_ID_BYTE_LENGTH = 16
-const UUID_REGEX = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/i
-const BYTE_TO_HEX: string[] = []
-for (let byte = 0; byte < 256; byte++) BYTE_TO_HEX.push(byte.toString(16).padStart(2, '0'))
-
-/** Canonical UUID string → 16 bytes. Returns `null` for anything else. */
-function uuidToBytes(uuid: string): Uint8Array | null {
-  if (!UUID_REGEX.test(uuid)) return null
-  const hex = uuid.split('-').join('')
-  const bytes = new Uint8Array(MEMBER_ID_BYTE_LENGTH)
-  for (let i = 0; i < MEMBER_ID_BYTE_LENGTH; i++) bytes[i] = Number.parseInt(hex.slice(i * 2, i * 2 + 2), 16)
-  return bytes
-}
-
-/** First 16 bytes → canonical lowercase UUID string. */
-function bytesToUuid(bytes: Uint8Array): string {
-  let uuid = ''
-  for (let i = 0; i < MEMBER_ID_BYTE_LENGTH; i++) {
-    if (i === 4 || i === 6 || i === 8 || i === 10) uuid += '-'
-    uuid += BYTE_TO_HEX[bytes[i]!]!
-  }
-  return uuid
-}
-
-/** Binary frame flags (one byte after the member ID). */
-const FRAME_FLAG_KEY = 0b0000_0001
-const FRAME_FLAG_TRACK = 0b0000_0010
-/** Track names stay tiny — they ride every frame. */
-const TRACK_MAX_BYTES = 64
-const frameTextEncoder = /* @__PURE__ */ new TextEncoder()
-const frameTextDecoder = /* @__PURE__ */ new TextDecoder()
-
-/** Binary relay format: `[16-byte member UUID][1-byte flags][?1-byte track length + track][payload]`.
- *  A plain publish costs one flag byte; named tracks (mic/camera/screen on one member lane) and
- *  the keyframe bit make media multiplexing first-class — no hand-rolled envelopes. */
-function frameWithMemberId(memberId: string, payload: Uint8Array, opts?: BinaryPublishOptions): Uint8Array {
-  const idBytes = uuidToBytes(memberId)
-  assert(idBytes, 'room member IDs are UUIDs')
-  let flags = opts?.keyFrame === true ? FRAME_FLAG_KEY : 0
-  let trackBytes: Uint8Array | null = null
-  if (opts?.track !== undefined) {
-    assertUsage(typeof opts.track === 'string' && opts.track.length > 0, 'track should be a non-empty string')
-    trackBytes = frameTextEncoder.encode(opts.track)
-    assertUsage(trackBytes.byteLength <= TRACK_MAX_BYTES, `track should be at most ${TRACK_MAX_BYTES} bytes`)
-    flags |= FRAME_FLAG_TRACK
-  }
-  const headerLength = MEMBER_ID_BYTE_LENGTH + 1 + (trackBytes ? 1 + trackBytes.byteLength : 0)
-  const framed = new Uint8Array(headerLength + payload.byteLength)
-  framed.set(idBytes, 0)
-  framed[MEMBER_ID_BYTE_LENGTH] = flags
-  if (trackBytes) {
-    framed[MEMBER_ID_BYTE_LENGTH + 1] = trackBytes.byteLength
-    framed.set(trackBytes, MEMBER_ID_BYTE_LENGTH + 2)
-  }
-  framed.set(payload, headerLength)
-  return framed
-}
-
-/** Split a binary relay frame into sender, track, keyframe bit, and payload. `null` on truncation. */
-function unframeMemberId(
-  data: Uint8Array,
-): { from: string; payload: Uint8Array; track: string | null; keyFrame: boolean } | null {
-  if (data.byteLength < MEMBER_ID_BYTE_LENGTH + 1) return null
-  const flags = data[MEMBER_ID_BYTE_LENGTH]!
-  const keyFrame = (flags & FRAME_FLAG_KEY) !== 0
-  let track: string | null = null
-  let offset = MEMBER_ID_BYTE_LENGTH + 1
-  if (flags & FRAME_FLAG_TRACK) {
-    if (data.byteLength < offset + 1) return null
-    const trackLength = data[offset]!
-    offset += 1
-    if (data.byteLength < offset + trackLength) return null
-    track = frameTextDecoder.decode(data.subarray(offset, offset + trackLength))
-    offset += trackLength
-  }
-  return { from: bytesToUuid(data), payload: data.subarray(offset), track, keyFrame }
-}
-
-// ---------------------------------------------------------------------------
-// RoomState — the local view of a room, driven by the event stream
-// ---------------------------------------------------------------------------
-
-type ListenerKind = 'data' | 'binary' | 'event' | 'activity'
-
-/** A binary listener's track filter: `undefined` = every track, `null` = the default lane only,
- *  a name = that track only. */
-type TrackFilter = string | null | undefined
-
-type MemberEntry = {
-  id: string
-  meta: ParticipantMeta
-  joinedAt: number
-  identity: string | null
-  /** Latest applied meta revision — stale and echoed `p-meta` events are absorbed. */
-  metaSeq: number
-  /** Named tracks the member is known to publish — grown by `track` events and rosters,
-   *  never shrunk (tracks live as long as the member). Drives all-track key subscriptions. */
-  tracks: Set<string>
-  remote: RemoteParticipant
-  dataCbs: Array<(data: unknown, info: ChannelPublishInfo) => unknown>
-  binaryCbs: Array<{
-    cb: (data: Uint8Array, info: ChannelPublishInfo & BinaryFrameInfo) => unknown
-    track: TrackFilter
-  }>
-  updateCbs: Array<(meta: ParticipantMeta, prev: ParticipantMeta) => void>
-  leaveCbs: Array<(cause?: LeaveCause) => void>
-}
-
-type RoomStateOptions = {
-  roomId: string
-  meta: RoomMeta
-  size: number
-  /** Either the authoritative roster, or just its size — a lazy view seeds with `{ count }`
-   *  and learns the members from its first `reconcile()` (KV read / streamed roster). */
-  seed: { members: MemberSnapshot[] } | { count: number }
-  /** The LWW stamp of the config `meta`/`size` were read from (see `applyRoomUpdate`). */
-  updateStamp: { at: number; by: string }
-  closed?: boolean
-  /** Fired whenever the number of attached listeners changes — lets the owner
-   *  (de)activate its event source (adapter subscription, wire subscription). */
-  onListenersChanged: () => void
-  /** A user callback threw — the owner decides how to report it. */
-  onCallbackError: (err: unknown) => void
-}
-
-/**
- * The local, event-driven view of a room: membership, metadata, and every user-facing callback.
- * Server and client share this class so event semantics are identical on both sides; only the
- * event *source* differs (adapter subscription vs relayed wire frames).
- *
- * Event application is idempotent — a `join` for a known member or a `leave` for an unknown one
- * is absorbed silently. This lets owners seed state from a snapshot and apply a concurrently
- * produced event stream without double-firing.
- */
-/** Stamped (non-enumerably) on every `RemoteParticipant` a `RoomState` mints — the backing
- *  that lets the serializer turn a view object back into (room, member) without a registry. */
-const ROOM_REMOTE_BRAND: unique symbol = Symbol.for('telefunc.RoomRemoteParticipant')
-
-type RemoteBacking = { state: RoomState; entry: MemberEntry }
-
-/** The `RoomState` backing of a minted `RemoteParticipant` — `null` for anything else. */
-function remoteBacking(value: unknown): RemoteBacking | null {
-  return typeof value === 'object' && value !== null && ROOM_REMOTE_BRAND in value
-    ? ((value as Record<typeof ROOM_REMOTE_BRAND, RemoteBacking>)[ROOM_REMOTE_BRAND] as RemoteBacking)
-    : null
-}
-
-class RoomState {
-  /** @internal — the owning `ServerRoom`/`ClientRoom`, for serialization backing. */
-  _owner: unknown = null
-  readonly roomId: string
-  meta: RoomMeta
-  size: number
-  closed: boolean
-  /** Bumped on every membership change — guards async KV reconciles against going stale. */
-  membershipVersion = 0
-  /** Bumped on every observable change (membership, participant meta, room config, closure) —
-   *  drives `onChange`/`snapshot()` cache invalidation. */
-  private _stateVersion = 0
-  private _snapshotCache: { version: number; value: RoomSnapshotView } | null = null
-  private readonly _changeCbs: Array<() => void> = []
-
-  private readonly _members = new Map<string, MemberEntry>()
-  private readonly _onListenersChanged: () => void
-  private readonly _onCallbackError: (err: unknown) => void
-
-  private readonly _roomDataCbs: Array<(data: unknown, info: ChannelPublishInfo, from: Sender) => unknown> = []
-  private readonly _roomBinaryCbs: Array<{
-    cb: (data: Uint8Array, info: ChannelPublishInfo & BinaryFrameInfo, from: Sender) => unknown
-    track: TrackFilter
-  }> = []
-  private readonly _joinCbs: Array<(member: RemoteParticipant) => void> = []
-  private readonly _leaveCbs: Array<(member: RemoteParticipant, cause?: LeaveCause) => void> = []
-  private readonly _updateCbs: Array<(meta: RoomMeta, prev: RoomMeta) => void> = []
-  private readonly _emptyCbs: Array<() => void> = []
-  private readonly _fullCbs: Array<() => void> = []
-  private readonly _closeCbs: Array<() => void> = []
-  private readonly _announceCbs: Array<(data: unknown, info: ChannelPublishInfo) => void> = []
-  private readonly _activityCbs: Array<(info: { timestamp: number }) => void> = []
-
-  private _eventListenerCount = 0
-  private _dataListenerCount = 0
-  private _binaryListenerCount = 0
-  private _activityListenerCount = 0
-  private _wasFull: boolean
-  private _updateStamp: { at: number; by: string }
-  private _rosterKnown: boolean
-  private _seedCount = 0
-
-  constructor(opts: RoomStateOptions) {
-    this.roomId = opts.roomId
-    this.meta = opts.meta
-    this.size = opts.size
-    this.closed = opts.closed === true
-    this._updateStamp = opts.updateStamp
-    this._onListenersChanged = opts.onListenersChanged
-    this._onCallbackError = opts.onCallbackError
-    if ('members' in opts.seed) {
-      this._rosterKnown = true
-      for (const member of opts.seed.members) this._createEntry(member)
-    } else {
-      this._rosterKnown = false
-      this._seedCount = opts.seed.count
-    }
-    this._wasFull = this.isFull
-  }
-
-  // ── Reads ──
-
-  /** Exact once the roster is known (seeded or reconciled); before that, the seeded count
-   *  adjusted by the events this view applied since. */
-  get count(): number {
-    return this._rosterKnown ? this._members.size : this._seedCount
-  }
-
-  /** Whether this view holds the authoritative member list (vs just a count). */
-  get rosterKnown(): boolean {
-    return this._rosterKnown
-  }
-
-  get isFull(): boolean {
-    return this.count >= this.size
-  }
-
-  /** Listeners needing the control stream (presence, lifecycle). */
-  get eventListenerCount(): number {
-    return this._eventListenerCount
-  }
-  /** Listeners needing the text data stream. */
-  get dataListenerCount(): number {
-    return this._dataListenerCount
-  }
-  /** Listeners needing the binary data stream. */
-  get binaryListenerCount(): number {
-    return this._binaryListenerCount
-  }
-  /** Listeners needing the activity trickle. */
-  get activityListenerCount(): number {
-    return this._activityListenerCount
-  }
-
-  /** Which (member, track) binary streams this holder needs delivered — drives the wire/adapter
-   *  subscriptions on both sides (client declares it, server aggregates it per stub). */
-  binaryWants(): BinaryWants {
-    const members: Record<string, TrackWants> = {}
-    for (const entry of this._members.values()) {
-      if (entry.binaryCbs.length > 0) members[entry.id] = trackWantsOf(entry.binaryCbs)
-    }
-    return { everyMember: trackWantsOf(this._roomBinaryCbs), members }
-  }
-
-  /** Named tracks the member is known to publish — `[]` for unknown members. */
-  memberTracks(id: string): string[] {
-    const entry = this._members.get(id)
-    return entry ? [...entry.tracks] : []
-  }
-
-  /** The text-lane twin of `binaryWants()`: `all` while room-level `subscribe()`rs exist,
-   *  otherwise exactly the members with participant-scoped listeners. */
-  textWants(): MemberWants {
-    if (this._roomDataCbs.length > 0) return { all: true, members: [] }
-    const members: string[] = []
-    for (const entry of this._members.values()) {
-      if (entry.dataCbs.length > 0) members.push(entry.id)
-    }
-    return { all: false, members }
-  }
-
-  getRemote(id: string): RemoteParticipant | null {
-    return this._members.get(id)?.remote ?? null
-  }
-
-  listRemotes(): RemoteParticipant[] {
-    return [...this._members.values()].map((entry) => entry.remote)
-  }
-
-  /** Member IDs currently known — drives isolated-mode per-member key subscriptions. */
-  listMemberIds(): string[] {
-    return [...this._members.keys()]
-  }
-
-  snapshotMembers(): MemberSnapshot[] {
-    return [...this._members.values()].map(({ id, meta, joinedAt, metaSeq, identity, tracks }) => ({
-      id,
-      meta,
-      joinedAt,
-      metaSeq,
-      identity,
-      ...(tracks.size === 0 ? {} : { tracks: [...tracks] }),
-    }))
-  }
-
-  /** Revival side of a serialized `RemoteParticipant`: the live view wins when it already knows
-   *  the member; otherwise the entry is seeded silently from the snapshot — no events fire, no
-   *  count adjustment (the seed count already included the member), and the streamed roster
-   *  reconciles it like any other pre-roster knowledge. */
-  ensureRemoteFromSnapshot(snap: MemberSnapshot): RemoteParticipant {
-    const existing = this._members.get(snap.id)
-    if (existing) return existing.remote
-    const remote = this._createEntry(snap).remote
-    this._bumpState()
-    return remote
-  }
-
-  // ── Listener registration (all return an unlisten function) ──
-
-  subscribe(cb: (data: unknown, info: ChannelPublishInfo, from: Sender) => unknown): () => void {
-    return this._register(this._roomDataCbs, cb, 'data')
-  }
-  subscribeBinary(
-    cb: (data: Uint8Array, info: ChannelPublishInfo & BinaryFrameInfo, from: Sender) => unknown,
-    opts?: { track?: string | null },
-  ): () => void {
-    return this._register(this._roomBinaryCbs, { cb, track: normalizeTrackFilter(opts) }, 'binary')
-  }
-  onJoin(cb: (member: RemoteParticipant) => void): () => void {
-    return this._register(this._joinCbs, cb, 'event')
-  }
-  onLeave(cb: (member: RemoteParticipant, cause?: LeaveCause) => void): () => void {
-    return this._register(this._leaveCbs, cb, 'event')
-  }
-  onUpdate(cb: (meta: RoomMeta, prev: RoomMeta) => void): () => void {
-    return this._register(this._updateCbs, cb, 'event')
-  }
-  onEmpty(cb: () => void): () => void {
-    return this._register(this._emptyCbs, cb, 'event')
-  }
-  onFull(cb: () => void): () => void {
-    return this._register(this._fullCbs, cb, 'event')
-  }
-  onClose(cb: () => void): () => void {
-    return this._register(this._closeCbs, cb, 'event')
-  }
-
-  onChange(cb: () => void): () => void {
-    return this._register(this._changeCbs, cb, 'event')
-  }
-
-  onActivity(cb: (info: { timestamp: number }) => void): () => void {
-    return this._register(this._activityCbs, cb, 'activity')
-  }
-
-  applyActivity(at: number): void {
-    this._fireAll(this._activityCbs, { timestamp: at })
-  }
-
-  /** A member published its first frame on a new named track (idempotent — echoes, rosters,
-   *  and the owner's local apply all land here). Unknown members are absorbed like any other
-   *  pre-roster event. */
-  applyTrack(id: string, track: string): void {
-    this._members.get(id)?.tracks.add(track)
-  }
-
-  /** Immutable view of the whole room — cached by state version, so the reference is stable
-   *  until something actually changes (the `useSyncExternalStore` contract). */
-  snapshot(): RoomSnapshotView {
-    if (this._snapshotCache?.version === this._stateVersion) return this._snapshotCache.value
-    const participants = Object.freeze(
-      [...this._members.values()].map(({ id, identity, meta, joinedAt }) =>
-        Object.freeze({ id, identity, meta, joinedAt }),
-      ),
-    )
-    const value = Object.freeze({
-      id: this.roomId,
-      meta: this.meta,
-      size: this.size,
-      count: this.count,
-      isClosed: this.closed,
-      participants,
-    })
-    this._snapshotCache = { version: this._stateVersion, value }
-    return value
-  }
-
-  /** State changed observably — invalidate the snapshot and tell `onChange` subscribers. */
-  private _bumpState(): void {
-    this._stateVersion++
-    this._fireAll(this._changeCbs)
-  }
-
-  /** Membership changed: guard async KV reconciles against going stale, and narrate the change. */
-  private _bumpMembership(): void {
-    this.membershipVersion++
-    this._bumpState()
-  }
-  onAnnounce(cb: (data: unknown, info: ChannelPublishInfo) => void): () => void {
-    return this._register(this._announceCbs, cb, 'event')
-  }
-
-  // ── Event application ──
-
-  applyJoin(id: string, meta: ParticipantMeta, joinedAt: number, identity?: string | null): void {
-    if (this.closed) return
-    const existing = this._members.get(id)
-    if (existing) {
-      // Echo of a member this side already applied (or got via snapshot) — absorb.
-      existing.meta = meta
-      existing.joinedAt = joinedAt
-      return
-    }
-    const entry = this._createEntry({ id, meta, joinedAt, metaSeq: 0, identity })
-    this._seedCount++ // pre-reconcile, `count` tracks the seed adjusted by applied events
-    this._bumpMembership()
-    this._fireAll(this._joinCbs, entry.remote)
-    this._checkFull()
-  }
-
-  applyLeave(id: string, cause?: LeaveCause): void {
-    const entry = this._members.get(id)
-    if (!entry) return // unknown here: absorbed (pre-reconcile misses correct at reconcile)
-    this._members.delete(id)
-    this._seedCount = Math.max(0, this._seedCount - 1)
-    this._bumpMembership()
-    this._fireAll(entry.leaveCbs, cause)
-    this._fireAll(this._leaveCbs, entry.remote, cause)
-    this._releaseEntryListeners(entry)
-    this._wasFull = this.isFull
-    if (this._members.size === 0) this._fireAll(this._emptyCbs)
-  }
-
-  /** Applies only revisions newer than the entry's — the origin's echo (same seq) and events
-   *  arriving behind a fresher reconcile are absorbed. */
-  applyParticipantMeta(id: string, meta: ParticipantMeta, prev: ParticipantMeta, seq: number): void {
-    const entry = this._members.get(id)
-    if (!entry || seq <= entry.metaSeq) return
-    entry.metaSeq = seq
-    entry.meta = meta
-    this._bumpState()
-    this._fireAll(entry.updateCbs, meta, prev)
-  }
-
-  /** Last-writer-wins by `(at, by)`: concurrent `Room.update()`s converge to the same winner on
-   *  every node regardless of arrival order, and the origin's echo (same stamp) is absorbed. */
-  applyRoomUpdate(meta: RoomMeta, prev: RoomMeta, size: number, at: number, by: string): void {
-    if (!stampNewer({ at, by }, this._updateStamp)) return
-    this._updateStamp = { at, by }
-    this.meta = meta
-    this.size = size
-    this._bumpState()
-    this._fireAll(this._updateCbs, meta, prev)
-    this._checkFull()
-  }
-
-  /** The stamp of the config this view currently reflects (serialized into room snapshots). */
-  get updateStamp(): { at: number; by: string } {
-    return this._updateStamp
-  }
-
-  /** Room closed: member-level cleanup callbacks run (decoders etc.), then `onClose`.
-   *  Room-level `onLeave`/`onEmpty` intentionally don't fire — `onClose` is the signal. */
-  applyClosed(cause: LeaveCause = { type: 'closed' }): void {
-    if (this.closed) return
-    this.closed = true
-    this._rosterKnown = true // authoritatively empty
-    this._bumpMembership()
-    for (const entry of this._members.values()) {
-      this._fireAll(entry.leaveCbs, cause)
-      this._releaseEntryListeners(entry)
-    }
-    this._members.clear()
-    this._fireAll(this._closeCbs)
-  }
-
-  applyAnnounce(data: unknown, info: ChannelPublishInfo): void {
-    this._fireAll(this._announceCbs, data, info)
-  }
-
-  /** Messages never wait on the roster: `from` is the live `RemoteParticipant` when this view
-   *  knows the sender, else the `{ id, meta }` snapshot the sender's node stamped into the
-   *  envelope. Control and data travel on separate lanes, so a message can beat its sender's
-   *  join — identity is in the message, delivery is immediate, and nothing drops. */
-  applyData(
-    from: string,
-    fromMeta: ParticipantMeta,
-    fromIdentity: string | null,
-    data: unknown,
-    info: ChannelPublishInfo,
-    suppress: boolean,
-  ): void {
-    if (suppress) return
-    const entry = this._members.get(from)
-    this._fireAll(this._roomDataCbs, data, info, entry?.remote ?? { id: from, meta: fromMeta, identity: fromIdentity })
-    if (entry) this._fireAll(entry.dataCbs, data, info)
-  }
-
-  /** Binary frames carry only the sender's ID — a pre-join frame surfaces as `{ id, meta: {} }`
-   *  (rare: binary pipelines attach per member via `onJoin`, so the roster is normally ahead).
-   *  `track`/`key` come from the frame header; listeners with a `track` filter receive only
-   *  that track's frames. */
-  applyBinary(
-    from: string,
-    payload: Uint8Array,
-    track: string | null,
-    keyFrame: boolean,
-    info: ChannelPublishInfo,
-    suppress: boolean,
-  ): void {
-    if (suppress) return
-    const frameInfo: ChannelPublishInfo & BinaryFrameInfo = { ...info, track, keyFrame }
-    const entry = this._members.get(from)
-    const sender = entry?.remote ?? { id: from, meta: {}, identity: null }
-    this._fireTrackFiltered(this._roomBinaryCbs, track, (cb) => cb(payload, frameInfo, sender))
-    if (entry) this._fireTrackFiltered(entry.binaryCbs, track, (cb) => cb(payload, frameInfo))
-  }
-
-  /** Fire the listeners whose track filter admits `track` (`undefined` = every track). */
-  private _fireTrackFiltered<CB>(
-    cbs: Array<{ cb: CB; track: TrackFilter }>,
-    track: string | null,
-    invoke: (cb: CB) => unknown,
-  ): void {
-    for (const { cb, track: want } of [...cbs]) {
-      if (want !== undefined && want !== track) continue
-      try {
-        invoke(cb)
-      } catch (err) {
-        this._onCallbackError(err)
-      }
-    }
-  }
-
-  /** Resync against an authoritative membership snapshot. The first reconcile is the roster
-   *  *load* and is silent (the documented pattern reads `getParticipants()` and then follows
-   *  events — narrating the load would double-render it). Every later reconcile is discovered
-   *  *drift*, and an observed view never mutates silently: the diff replays through the normal
-   *  appliers, so `onJoin`/`onLeave`/`onUpdate` fire exactly once per real change — whichever
-   *  path (event or reconcile) learns of it first, the other is absorbed as an echo. Returns
-   *  whether anything drifted, so the owner can re-sync downstream views (client stubs). */
-  reconcile(members: MemberSnapshot[]): boolean {
-    if (!this._rosterKnown) {
-      // First load: silent — but entries can already exist (pre-roster join events, revived
-      // views). Those objects must survive: listeners hang off them and revived handles must
-      // stay `===` with the view. Keep the object, refresh the facts; entries the authoritative
-      // roster doesn't know left before the load — their leave is narrated like any other.
-      this._rosterKnown = true
-      this._bumpMembership()
-      const seen = new Set<string>()
-      for (const member of members) {
-        seen.add(member.id)
-        const existing = this._members.get(member.id)
-        if (!existing) {
-          this._createEntry(member)
-          continue
-        }
-        if (member.metaSeq > existing.metaSeq) {
-          existing.meta = member.meta
-          existing.metaSeq = member.metaSeq
-        }
-        existing.joinedAt = member.joinedAt
-        for (const track of member.tracks ?? []) existing.tracks.add(track)
-      }
-      for (const id of [...this._members.keys()]) {
-        if (!seen.has(id)) this.applyLeave(id)
-      }
-      this._wasFull = this.isFull
-      return false
-    }
-
-    this._bumpMembership()
-    let drifted = false
-    const seen = new Set<string>()
-    for (const member of members) {
-      seen.add(member.id)
-      const entry = this._members.get(member.id)
-      if (!entry) {
-        this.applyJoin(member.id, member.meta, member.joinedAt, member.identity)
-        const created = this._members.get(member.id)
-        if (created) {
-          created.metaSeq = member.metaSeq
-          for (const track of member.tracks ?? []) created.tracks.add(track)
-        }
-        drifted = true
-      } else {
-        if (member.metaSeq > entry.metaSeq) {
-          this.applyParticipantMeta(member.id, member.meta, entry.meta, member.metaSeq)
-          drifted = true
-        }
-        entry.joinedAt = member.joinedAt
-        for (const track of member.tracks ?? []) entry.tracks.add(track)
-      }
-    }
-    for (const id of [...this._members.keys()]) {
-      if (!seen.has(id)) {
-        this.applyLeave(id)
-        drifted = true
-      }
-    }
-    return drifted
-  }
-
-  // ── Private ──
-
-  private _checkFull(): void {
-    const full = this.isFull
-    if (full && !this._wasFull) this._fireAll(this._fullCbs)
-    this._wasFull = full
-  }
-
-  private _createEntry(entrySeed: MemberSnapshot): MemberEntry {
-    const { id, meta, joinedAt } = entrySeed
-    const entry: MemberEntry = {
-      id,
-      meta,
-      joinedAt,
-      identity: entrySeed.identity ?? null,
-      metaSeq: entrySeed.metaSeq,
-      tracks: new Set(entrySeed.tracks),
-      remote: {
-        id,
-        get meta() {
-          return entry.meta
-        },
-        get joinedAt() {
-          return entry.joinedAt
-        },
-        get identity() {
-          return entry.identity
-        },
-        subscribe: (cb) => this._register(entry.dataCbs, cb, 'data'),
-        subscribeBinary: (cb, opts) =>
-          this._register(entry.binaryCbs, { cb, track: normalizeTrackFilter(opts) }, 'binary'),
-        onUpdate: (cb) => this._register(entry.updateCbs, cb, 'event'),
-        onLeave: (cb: (cause?: LeaveCause) => void) => this._register(entry.leaveCbs, cb, 'event'),
-      },
-      dataCbs: [],
-      binaryCbs: [],
-      updateCbs: [],
-      leaveCbs: [],
-    }
-    // Non-enumerable: never serialized as data, invisible to user iteration.
-    Object.defineProperty(entry.remote, ROOM_REMOTE_BRAND, { value: { state: this, entry } satisfies RemoteBacking })
-    this._members.set(id, entry)
-    return entry
-  }
-
-  private _register<T>(list: T[], cb: T, kind: ListenerKind): () => void {
-    list.push(cb)
-    this._bumpListenerCount(kind, 1)
-    // List membership is the source of truth: `_releaseEntryListeners` may have already
-    // emptied the list, and a second unlisten call must not decrement twice.
-    return () => {
-      const i = list.indexOf(cb)
-      if (i < 0) return
-      list.splice(i, 1)
-      this._bumpListenerCount(kind, -1)
-    }
-  }
-
-  /** A member entry is being discarded — its listeners die with it. Releasing them keeps the
-   *  counters truthful (callers rarely unsubscribe in `onLeave`), which lets the owners drop
-   *  wire/adapter subscriptions the departed member was holding open. */
-  private _releaseEntryListeners(entry: MemberEntry): void {
-    this._bumpListenerCount('data', -entry.dataCbs.length)
-    this._bumpListenerCount('binary', -entry.binaryCbs.length)
-    this._bumpListenerCount('event', -(entry.updateCbs.length + entry.leaveCbs.length))
-    entry.dataCbs.length = 0
-    entry.binaryCbs.length = 0
-    entry.updateCbs.length = 0
-    entry.leaveCbs.length = 0
-  }
-
-  private _bumpListenerCount(kind: ListenerKind, delta: number): void {
-    if (kind === 'data') this._dataListenerCount += delta
-    else if (kind === 'binary') this._binaryListenerCount += delta
-    else if (kind === 'activity') this._activityListenerCount += delta
-    else this._eventListenerCount += delta
-    this._onListenersChanged()
-  }
-
-  private _fireAll<Args extends unknown[]>(cbs: Array<(...args: Args) => unknown>, ...args: Args): void {
-    for (const cb of [...cbs]) {
-      try {
-        cb(...args)
-      } catch (err) {
-        this._onCallbackError(err)
-      }
-    }
-  }
-}
-
-/** Validate a `subscribeBinary` track option: `undefined` = every track, `null` = the default
- *  lane, a non-empty name = that track. */
-function normalizeTrackFilter(opts: { track?: string | null } | undefined): TrackFilter {
-  const track = opts?.track
-  if (track === undefined || track === null) return track
-  assertUsage(typeof track === 'string' && track.length > 0, 'subscribeBinary() track should be a non-empty string')
-  return track
-}
-
-/** Fold a listener list's track filters into the `TrackWants` they add up to. */
-function trackWantsOf(cbs: ReadonlyArray<{ track: TrackFilter }>): TrackWants {
-  const wants = emptyTrackWants()
-  for (const { track } of cbs) {
-    if (track === undefined) return { all: true, tracks: [] }
-    const asTrack = track ?? DEFAULT_TRACK
-    if (!wants.tracks.includes(asTrack)) wants.tracks.push(asTrack)
-  }
-  return wants
-}
-
-// ---------------------------------------------------------------------------
-// ParticipantBase — the shared half of every LocalParticipant
-// ---------------------------------------------------------------------------
-
-/**
- * The private-message inbox and the leave lifecycle, identical on server and client.
- * Flavors supply the transport through the abstract operations and their own error pipeline.
- */
-/** A delivered private message, as stamped by the sender's node. */
-type InboxMessage = {
-  from: string
-  fromMeta: ParticipantMeta | null
-  fromIdentity: string | null
-  data: unknown
-}
-
-/** Pre-listen inbox hold: count-capped, drop-oldest. The DM lane is the only
- *  unconditionally-delivered lane (addressed — there are no wants to gate it on), so it's the
- *  only lane with a client-side attach window to bridge; every room lane is want-gated at the
- *  server and has nothing to hold. Message size is bounded upstream by the wire-protocol
- *  `messageLimit`, which bounds the hold's memory too. */
-const PENDING_INBOX_MAX_COUNT = 64
-
-abstract class ParticipantBase implements LocalParticipant {
-  readonly id: string
-  readonly identity: string | null
-  readonly selfDelivery: boolean
-  /** @internal */ _meta: ParticipantMeta
-  protected _left = false
-  private _leftCause: LeaveCause | null = null
-  private _leaveCbs: Array<(cause: LeaveCause) => void> = []
-  private readonly _messageCbs: Array<(data: unknown, from: Sender | null) => void> = []
-  /** DMs delivered before the first `listen()` — held bounded, flushed on attach, then never
-   *  allocated again (`null` = flushed or empty; zero steady-state cost). */
-  private _pendingInbox: InboxMessage[] | null = null
-
-  constructor(id: string, meta: ParticipantMeta, selfDelivery: boolean, identity: string | null) {
-    this.id = id
-    this._meta = meta
-    this.selfDelivery = selfDelivery
-    this.identity = identity
-  }
-
-  get meta(): ParticipantMeta {
-    return this._meta
-  }
-
-  abstract publish(data: unknown): Promise<ChannelPublishAck>
-  abstract publishBinary(data: Uint8Array, options?: BinaryPublishOptions): Promise<ChannelPublishAck>
-  abstract send(to: string | Sender, data: unknown): Promise<void>
-  abstract setMeta(meta: ParticipantMeta): Promise<void>
-  abstract leave(): Promise<void>
-  /** A user callback threw — each side reports through its own pipeline. */
-  protected abstract _reportError(err: unknown): void
-
-  listen(callback: (data: unknown, from: Sender | null) => void): () => void {
-    this._messageCbs.push(callback)
-    if (this._pendingInbox) {
-      const held = this._pendingInbox
-      this._pendingInbox = null
-      for (const msg of held) this._deliverMessage(msg)
-    }
-    return () => {
-      const i = this._messageCbs.indexOf(callback)
-      if (i >= 0) this._messageCbs.splice(i, 1)
-    }
-  }
-
-  /** @internal — a direct message arrived on this member's inbox. `from`/`fromMeta` come from
-   *  the wire envelope; `resolve` upgrades to the live `RemoteParticipant` when a room view
-   *  exists. An empty `from` is the wire encoding of a room-authored message → `null`. */
-  _deliverMessage(msg: InboxMessage): void {
-    if (this._messageCbs.length === 0) {
-      // A reactive send can beat the app's `listen()` by a tick — hold it (bounded) instead of
-      // dropping it. A departed participant will never flush: drop.
-      if (this._left) return
-      const pending = (this._pendingInbox ??= [])
-      pending.push(msg)
-      if (pending.length > PENDING_INBOX_MAX_COUNT) pending.shift()
-      return
-    }
-    const { from, fromMeta, fromIdentity, data } = msg
-    const sender =
-      from === '' ? null : (this._resolveSender(from) ?? { id: from, meta: fromMeta ?? {}, identity: fromIdentity })
-    for (const cb of [...this._messageCbs]) {
-      try {
-        cb(data, sender)
-      } catch (err) {
-        this._reportError(err)
-      }
-    }
-  }
-
-  /** The live room-backed sender, when this flavor has a room view. */
-  protected _resolveSender(_id: string): Sender | null {
-    return null
-  }
-
-  onLeave(callback: (cause: LeaveCause) => void): () => void {
-    if (this._leftCause) {
-      this._invoke(callback, this._leftCause)
-      return () => {}
-    }
-    this._leaveCbs.push(callback)
-    return () => {
-      const i = this._leaveCbs.indexOf(callback)
-      if (i >= 0) this._leaveCbs.splice(i, 1)
-    }
-  }
-
-  /** @internal — the member is gone; `cause` says how. A local participant always knows its
-   *  cause: its holder either initiated the leave or witnessed the event/closure that caused it. */
-  _onLeft(cause: LeaveCause): void {
-    this._left = true
-    if (this._leftCause) return
-    this._leftCause = cause
-    this._pendingInbox = null
-    const cbs = this._leaveCbs
-    this._leaveCbs = []
-    for (const cb of cbs) this._invoke(cb, cause)
-  }
-
-  protected _assertActive(): void {
-    if (this._left) throw new Error('Participant has left the room')
-  }
-
-  private _invoke(cb: (cause: LeaveCause) => void, cause: LeaveCause): void {
-    try {
-      cb(cause)
-    } catch (err) {
-      this._reportError(err)
-    }
-  }
-}
