+++ packages/telefunc/wire-protocol/types.ts
@@ -48,7 +48,7 @@ import type { ServerBroadcast } from './server/server-broadcast.js'
 import type { ClientChannel, ClientBroadcast } from './client/channel.js'
 import type { ServerRoom, ServerLocalParticipant } from './room/server.js'
 import type { Room, LocalParticipant, RemoteParticipant } from './room/types.js'
-import type { RoomSnapshotMetadata, ParticipantStubMetadata } from './room/shared.js'
+import type { RoomSnapshotMetadata, ParticipantStubMetadata } from './room/protocol.js'
 import type { AbortError } from '../shared/Abort.js'
 import type { ShieldValidators } from '../node/server/shield.js'
 import type { FileDownload, BlobDownload } from './client/response/DownloadClasses.js'
